Converted map2cam to cpp / functions
